Bug#35000554: assertion error in EstimateAggregateNoRollupRows()
When using hypergraph, queries like:

SELECT ... FROM <table> GROUP BY <field 1>..<field N>;

where one or more of <field 1>..<field N> is NULL in all rows and also
has a histogram, we got a row estimate of 0 (and an assertion error in
debug mode). This commit fixes that bug by also counting NULL when
estimating the number of distinct values (and hence the number of
result rows).

This commit also fixes a related problem: when we use histograms to
estimate the selectivity for a term like 'field1=field2', we get too high
selectivity (and therefore too high a row estimate) if the field we
look at has NULL values.

Finally, this patch adds checks so that we ignore histograms that were built
on empty tables and calculate estimates using fallback methods in such cases.

Change-Id: Ibb9eeff1845defaeb2b31b81d957997e94c09d50