Bug#35089015 Duplicate key in tables.engine during DD table install
Before NDB table definitions can be stored in the data dictionary,
fruit of metadata synchronization between NDB and the ndbcluster
plugin, any existing definition must be removed. Table definitions
have two unique values, the table name and the
ndbcluster-<se_private_id> pair. Hence, the install code checks if
there is any existing definition with the same table name and removes
it. It then proceeds to check if the removed table and the one being
installed have the same se_private_id. If they do not, the definition
that is occupying the se_private_id is removed too (considered stale).

But if no existing definition was found by the search using the
table's name then no definition will be dropped even if one is
occupying se_private_id. This will lead to a duplicate key error when
storing the new table. The code at the "store_table" function attempts
to clear the diagnostics area, remove the stale definition of
se_private_id and try to store it once again. However, in a release
build, the diagnostics area is not cleared. The error is thus leaked
and will be presented to the user.

To fix this, this commit removes any stale table definition. It
renders the attempt of "store_table" useless, but it is left out for
the time being to safeguard.

Change-Id: I9fb1d5e02445781f8f3e7c5291fbe10b73beb59b