Bug#34973220 Correlated scalar subquery transf. to derived table: wrong result dup predicate
We had missed resetting the status as outer reference of the
occurrences 2..n of the same outer field.  The subquery had more than
one reference to the same outer field, but while collecting them we
conflated the items. This lead to us clearing the outer status of only
one of the Item_field's.

The solution is to collect all occurrences of Item_field which are
outer references.

Change-Id: Icd4cebffa7f010bb9e0af3e1e13d76d5302004ab