Bug#35067083 Router compatilibilty issue with old Shell
[WL#15086]

When the cluster metadata version is 1.x.x the Router tries to use
mysql_innodb_cluster_metadata.v2_routers view which does not exist
in that metadata version.

If the metadata version is 2.2.x there is a new view for the Router
options v2_router_options. For the lower versions, the Router was
falling back to using v2_routers, which did not exist in 1.x.x
version. This patch removes querying for the Router options for
pre-2.2.x metadata versions. The Router options were never used nor
queried in the previous versions and the only option that the Router
now handles is read_replicas_policy which requires 2.2.x version.

Change-Id: Ic31fd1b619977c07e6758037064993e5f10fddc0