Bug#35068222: Incorrect condition in static_assert in SimulateJoin()
SimulateJoin() has some disabled code for using the smaller table as
build table when simulating a hash inner join. It is disabled because
the estimated cost of a hash join happens to be the same regardless of
the join order with the current cost constants. There is a
static_assert that is supposed notify us if the code needs to be
enabled, but that static_assert compares the hash build cost with
itself, so it will never fail.

This patch changes the static_assert so that it compares the build
cost with the probe cost, as originally intended. And it changes the
code so that it uses the smaller table, not the table with the lowest
cost, as the build table. (The cost of reading the tables has no
relevance to the join order, as the full cost of reading both tables
needs to be paid regardless of in which order the tables are read. It
is the number of rows in each table that affects how many times the
build-one-row cost and the probe-one-row cost have to be paid.) It is
also extended to handle all commutative joins, not only inner joins,
in preparation for full outer join.

There is no fundamental reason why the build and probe cost factors
should be equal. Quite the opposite, actually. Measurements suggest
the build factor should be considerably higher than the probe factor.
The patch therefore also enables the code. Even though this is
unnecessary when the two cost factors happen to be equal and it
affects performance in some cases, it has to be done at some point.
Doing it early is better, so that it gets some testing and so that the
performance hit doesn't show up as a surprise in the performance
profiles once the cost model constants are calibrated.

The microbenchmarks in graph_simplification-t.cc show that the graph
simplification takes approximately 10% more time for the clique join
test cases. The star join test cases take approximately 5% more time
to simplify.

Change-Id: If50b2fa831d996341f7d181c3fd811ab46a26a75