Bug#34801210 Error Illegal mix of collations when using views and IN statements
During condition optimization, all eq conditions are
analyzed in check_simple_equality() including conditions
from view expansions. If there is more than one field=const
conditions on the same field, the const values are compared
to make sure its a duplicate and not an impossible cluse.
In this bug scenario, the const comparison leads to an
'Illegal mix of collations' error due to different charsets
of the const values ('utf8mb4_sv_0900_ai_ci' from view and
'utf8mb4_general_ci' from the seelct in current session).
This error happens only with charsets that do not have
contractions. When using charsets with contractions, the
condition merge is not attempted and const comparison is
avoided.

Solution:
If the charsets of both const are not comparable, skip the
condition merge attempt and handle it like the case with
contracting charsets.

Change-Id: Ic3064fcfe13e4ad942b7d8b6864114082e920912