Bug#35071435 Repetitive code in the Routing plugin
MySQLRouting and MySQLRoutingContext classes both take around 20 parameters.
Most of them are 1:1 copies of the routing plugin configuration structure.
It does not scale too well, adding an option to the plugin requires also
adding it to the 2 constructors.

Solution:
Pass the RoutingConfig struct as a parameter to the constructor
of MySQLRouting and MySQLRoutingContext instead of a separate config
options.

Change-Id: If4291b1a879394d3ae024e99a0c66831125c4f3b