Bug 34823988 - CopyFrag waiting for a lock not closed on TC Takeover, stalls TC Takeover
The running node from a node group scans each fragment (copyFrag) and
sends the rows to the starting peer in order to make it
synchronized. If a row from the fragment is locked exclusively by a
user transaction, it will block the scan from reading it and the
copyFrag stalls.

If the starting node fails during the CopyFrag phase then the normal
node failure handling will occur. The cordinator node's TC will
perform TC takeover of the user transactions from the TCs on the
failing node. Since the scan that aid copying the fragment data over
to the starting node is 'internal', it is not a candidate for
takeover. Thus the takeover TC will mark the copyFrag scan to be
closed at the next opportunity, and wait until it is closed.

Problem: If the CopyFrag scan is in the 'waiting for row lock' state,
the close of the marked scan is not performed.  Consequences are : TC
takeover stalls waiting for the close, causing unfinished node failure
handling, GCP stall affecting Redo, LCP and Replication.

Fix: Close the marked CopyFrag scan if a node failure occurs while
copyFrag is waiting for a row lock.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>

Change-Id: Idd9971ef4655368d80a9e819270351323ae38805