Bug#34999683 Auth OCI plugin options are not copied
Problem:
1) When setting plugin option, either oci-config-file or
authentication-oci-client-config-profile, the char* passed is not
copied, but only saved the pointer.
With this, user variable has to live until the end of the plugin
deinit.

2) After one sets a custom option value for OCI authentication
plugin, there is no way to get the option back to its default
value and the same is used for all connections going forward.

Solution:
1) The plugin options are now copied and the same is released in
plugin deinit.

2) One can now reset values to default by passing nullptr.

Change-Id: Ia3fef15cb4714798dac7e5bd5397c66fd1f9f6f8