Bug#34426943: wrong default value for GEOMETRY column
Problem: When creating a table from a SELECT which has an expression
of GEOMETRY type, MySQL generates an empty string as DEFAULT.
Solution: Since GEOMETRY is a BLOB based data type, and such types
do not allow default values, the fix is to not allow DEFAULT
values to such columns.

Patch was created based on a contribution by Tencent.

Change-Id: Iaf1450c0fc231956248614a52b09c9be4ebfa805