BUG#35083373: mysqlbinlog ignores errors in files having LOG_EVENT_BINLOG_IN_USE_F set
* Background

When mysqlbinlog encounters any error while reading an event, it
normally writes an error message and returns a nonzero exit code. That
is good.

* Problem

However, for the active binary log file (or any binary log where the
format_description_log_event has the LOG_EVENT_BINLOG_IN_USE_F flag
set), it does not write a message, and returns exit code 0. So it
hides the error.

* Analysis

The server sets the IN_USE flag when it creates a binary log file, and
clears it when rotating the binary log, when shutting down cleanly,
and after successful binary log recovery. The flag is only set in
binary logs, not in relay logs. Therefore, the flag is set in the
following cases:

 1. We are viewing the "active" binary log; the one the server
    currently writes transactions to.

 2. We are viewing a binary log that was active, but then the server
    crashed and has not yet recovered.

 3. We are viewing a binary log that was active, but then the server
    crashed and recovery failed.

In all cases, it is possible that the file is truncated just because
we observe it when the server has only half-completed a write
operation.  So when mysqlbinlog finds that the file is truncated:

- In case 2 and 3, the situation is a bit exceptional so reporting an
error is probably helpful.

- Case 1 can possibly be a normal scenario in some use case (e.g. a
user running mysqlbinlog on the active binary log as part of a backup
operation).

mysqlbinlog cannot know if case 1, 2 or 3 applies, so to support the
possible use cases related to case 1, we should keep the suppression.
However:

- It currently suppresses all errors. It would be better to suppress
only errors related to truncated events.

- It would be helpful to print an informational message as a
comment to the output.

* Fix

Mysqlbinlog:

- Suppress only errors related to truncated events.

- When suppressing an error, print a comment instead.

- Update the help text for --force-if-open

Tests:

- Wrote binlog_corrupted_event_mysqlbinlog to test this bug.

- Made assert_command_output.inc print the command in all error cases.

- Moved copy_file_segment.inc from extra/binlog_tests (which is for
  test-specific include files) to include (which is for generic include
  files) and expanded a little bit on the usage instructions.
  Adjusted binlog_open_errors.test accordingly.

Change-Id: If01cf5d37df14a1a477728dddcf71cf6ed436634