Bug#34862188: Assertion failures in hypergraph when optimizer_max_subgraph_pairs is low
When setting optimizer_max_subgraph_pairs very low (for example to 1)
to reduce the search space as much as possible, the hypergraph
optimizer often fails to come up with a plan for queries that access
the INFORMATION_SCHEMA.STATISTICS view, which includes the SHOW
INDEXES and SHOW KEYS commands.

The problem is that graph simplification uses a quick but unreliable
check for cycles in the join dependency graph in order to reduce the
number of times it has to do a slower check to verify that the
simplified graph is joinable.

The quick check is known to give false negatives, so negatives are
always verified with the slow check. As we found out in bug#33620355,
it can also give false positives when the hypergraph contains cycles.
The problem in this bug is again a false positive. The join dependency
graph says that the "best" simplification step to apply next will
introduce a cycle in the dependency graph, and make the graph
unjoinable. It therefore adds the opposite restriction to the graph,
unconditionally, to ensure that this simplification step is not tried
again. But, in reality, the suggested simplification step would not
have made the graph unjoinable. However, adding the opposite
restriction does make it unjoinable.

Fixed by using the slow check (GraphIsJoinable()) to verify that the
opposite step won't make the graph unjoinable, and if it does, revert
the change to the hypergraph and instead remove that candidate step
forcefully from the priority queue using the code added for
bug#33620355.

"mtr --hypergraph --mysqld=--optimizer-max-subgraph_pairs=1" now runs
cleanly, apart from some differences in EXPLAIN output (expected) and
benign result differences (row ordering in unordered queries).

Using the slower check more frequently does not affect the
microbenchmarks in graph_simplification-t.cc negatively, as the
situation where the extra check is needed is relatively rare.

Change-Id: Ie1b6a156f842802e6963543dad6139561979e5b2