BUG#35068274 - There is no local IP address matching the one configured for the local node
When the MySQL instance was being started by Systemd, no IP
had yet been assigned by DHCP. Journalctl shows the MySQL server was started
after the network target but before the network-online target after which an
IP was available.

This fix is to ensure that all configured network devices are up and have an
IP address assigned before the service is started.

Change-Id: If360a31004b76136cb86f416de41452c075f242a