Bug#35101630 Correlated scalar subquery transf. to derived table:              allows unsupported expr
We allowed for transformation of an equality expression of the form

  <non-corr. column> = <correlated column> + <non-corr. column>

or more generally, the right side was accepted as eligible for
transformation as long as t contained one column that was
correlated. But the above is equivalent to

  <non-corr. column> - <non-corr. column> = <correlated column>

for which we already have a check leading to rejection of transform,
Cf. WL#13520 requirement number nine:

"FR#9. In an eligible WHERE clause predicate, each operand should be a
       simple column reference"

Given
        <non-correlated column> = <correlated expression>

this patch adds a check that the <correlated expression> doesn't contain
a non-correlated column.

With the patch, a RAPID error is seen; an offload no longer happens,
from a query in subquery_sj_{varchar,dict}:

SELECT ( SELECT a FROM t1 AS t2
         WHERE a IN (SELECT t1.a+t2.a FROM t1 AS t3) )
FROM t1 AS t1

-> is semijoin transformed to

SELECT ( SELECT t2.a
         FROM t1 t2
         SEMI JOIN
         (t1 t3)
         WHERE ((t2.a = (t1.a + t2.a)))) AS `....`
FROM t1

This used to be transformed further [1] due to the presence of
bug#35101630 but now gets blocked because of the right hand side
(correlated side) has "+ t2.a" (non-correlated).

[1]:
select `derived_1_2`.`a` AS `...`
from `mysql`.`t1`
   left join
     ( select `mysql`.`t2`.`a` AS `a`,
             count(0) AS `Name_exp_2`
       from `mysql`.`t1` `t2`
            semi join
            (`mysql`.`t1` `t3`)
       group by `mysql`.`t2`.`a`) `derived_1_2`
   on (((`derived_1_2`.`a` = (`mysql`.`t1`.`a` + `derived_1_2`.`a`)) and
        reject_if((`derived_1_2`.`Name_exp_2` > 1)))) where true

Were the equality predicate transformed from

     t2.a = (t1.a + t2.a)
to
     t2.a - t2.a = t1.a

it would be rejected also (function), but allowed by WL#15540 (we would
add GROUP ON (t2.a - t2.a), or GROUP ON (0-0/*extra zero to avoid
unknown column '0'*/).

Change-Id: Ia357d09646d5decd1d9593e4d5a85f230f6cb774
Change-Id: Ia357d09646d5decd1d9593e4d5a85f230f6cb774