Bug#35084378: Client authentication plugins cannot be loaded
The fix for bug 35046862 undefined MYSQL_DYNAMIC_PLUGIN for client
plugins.
This is correct, but the macros in client_plugin.h need this to
declare certain symbols on windows as exported so that the dll
is a proper client dynamic plugin.
Added a new macro MYSQL_DYNAMIC_CLIENT_PLUGIN that is set for
all dynamically loadable client plugins.
Made sure client plugins depend on nothing but the new macro.
Renamed the client export macro to its own unique name to
support code bases that are included in both the client and the
server.
Removed an undef from client_plugin now that the macros are different.

Change-Id: I936118f84935006793398fa7fa0e402ce4eae579