Bug#35106292 NDB API should allow to setValue() of a PK column to the same or equal value
The NDB API currently prevents using ::setValue() to assign any
values to a primary key column by returning error code 4202:

  4202: "Set value on tuple key attribute is not allowed"

Even assigning a value identical to the existing one is rejected with such
an error. This is too strict as we should allow primary key column to
be directly 'updated' to a value 'equal by comparison rules' value.

Note that when eg. a case insensitive collation is used for a PK column,
the values 'xyz' and 'XYZ' are 'equal'

The NdbRecord based interface has allowed such 'equal' updates since it was
introduced by instead rellying on the data nodes to check for updates
to non-equal values, and return error 897 ('Update attempt of primary key
via ndbcluster internal api') if such updates are found.

The exact same checks are performed at the data nodes for updates having
its origin in the setValue (and family) instructions. Thus it should be
safe to entirely remove the too struct PK update ckecks from THE NDB API.

Note that in addition to setValue(), there is a family of similar operations
where a column value can be set or updated.
(Write_attr(), incValue(), subValue(), ...)

Patch remove the error 4202 error check and returns from the NDB API
and leave it to the data nodes to return error 897 where a PK update
violation is found. This is exactly the same way as handled by the
NdbRecord interface.

The test case 'testNdbApi -n NdbPKUpdateWithSetValue' is added as well.

Change-Id: I57d7fc98d9a0247b5f592597bf1e399b7259f7d5