Bug#34388068 NDB: Update to char data type columns in primay keys are lost on replica
Changes to character data type column values might not be performed
on the Mysql replicated slaves if:

  - The column value is changed to a value being binary different from the
    'before'-value, but still 'equal' according to the comparison rules
    existing for the specific data type / collation.
    E.g.
     -- Trailing spaces removed in a varchar
     -- Change between upper/lower case in a case insensitive collation.

  AND

  - All columns in the specific table are declared as being part of
    the primary key.

Root cause for this bug is that such changes to any primary key values
were not contained in UPDATE triggers and the NdbEventsOperations used to
propogate changes into the binlog, such that the replica could be updated.
There seems to have been an underlying assumption that primary keys
are 'constant', thus the NdbEventOperations API were not able to receive
changes in the primary key values when unpacking the received
BEFORE/AFTER values. Neither did the TUP block include and primary key
changes in an UPDATE trigger - Actually it explicitely eliminated the
primary key attributes from the 'changeMask' of attributes being changed.

Patch consist of 2 changes:

1) DbtupTrigger sets up the mask of attributes to monitor, and later include
in the BEFORE/AFTER values in an update trigger. It used to make the asumption
that PK values were 'constant', such that it explicitely eliminated the
PK attribute from the 'attrbiteMask' to monitor (and send). For an UPDATE
trigger we now include the PK attributes as well and include them in the
BEFORE values in the update trigger if some key values actually changed.
(Note that the UPDATE-AFTER values are always reflected in the KEY)

2) NdbEventOperationImpl::receive_event() receive the BEFORE and AFTER
image in an update-event. It would skip over any attribute values
received where a NdbRecAttr was not found to receive the value into,
which also included any PK values sent as part of the BEFORE/AFTER
values. This code has now been enhanced to also receive PK values
being part of the BEFORE values.

Online upgrade considerations:
As the existing BEFORE/AFTER-receive_event() code just skipped over attributes
not being expected. Thus sending BEFORE_PK values from a data node with
the modfied DbtupTrigger code will only result in the PK being ignored.
( -> Behave just as before)

The other way around will be just the same. Thus it should be perfectly
safe to run with a mix of (non-)upgraded clients and data nodes.

Other changes in patch:

Introduced the pkAttributeMask which is a bitmap of the PK columns in a table.
That made the method Dbtup::primaryKey() obsolete and simplified some code
where we previously had to loop over all attributes in a table definition
whenever include/exclude of the PK columns were needed.

Also simplified the code setting up the 'numAttrsToRead' in Dbtup::readTriggerInfo().
Calls to setAttrIds() is moved out of a composite if .. else code block
to a common place later.

The testcase 'test_event -n PrimaryKeyUpdates' has been added as part of this
fix, as well as extensions to the MTR test cases
ndb_rpl_cmp_varchar_charset.test and ndb_rpl_rowsize30000.test.

Change-Id: Ic5bb6dca80aa94a8cca4570c6d87c6c1639b5387
(cherry picked from commit 15d0a148071bc53f6984d3f031b723ae21a13595)