Bug#35112431 mysqli_real_connect() hangs with PASSTHROUGH
When connecting with

- PHP (mysqli/pdo)
- no encryption (the default for PHP)
- to a caching_sha2_password account
- after authentication the same account succeeded before
- and router with PASSTHROUGH/AS_CLIENT or PREFERRED/AS_CLIENT

the mysqli_real_connect() hangs.

Background
----------

The server sends:

- 0x01 0x03 -- fast-auth-ok
- 0x00 ...  -- ok

PHP's mysqlnd_wireprotocol.cc expects

- 0x01 0x03 -- fast-auth-ok
- 0x00 ...  -- ok

but the router only sends:

- 0x00 ...  -- ok

libmysqlclient accepts the router's shortend packet flow, but PHP doesn't
and waits for the right packet.

Change
------

Only surpress the "fast-auth-ok" packet from the server, if the client
has not sent the password to the router.

Change-Id: Ia56958d67084a81fcdab0dc1a592586d3a980a76