Bug#34668756 - impossible on not optimized in 8.x like 5.6
Issue:
8.0 behaves differently than 5.6 when we have an always false join
condition for some outer join operations.

Analysis:
For a query with an outer join having a join condition with many
predicates along with an always false predicate, makes optimizer in
8.0 not mark the inner table of the outer join as const. This
creates an unnecessary join and the inner table gets scanned.
In 5.7, optimizer would mark the inner table as const with the null
row flag set. This results in scanning only the outer table with inner
table fields always returning NULLs.
The change in plan in 8.0 is because of the early detection of the
always false clause i.e 8.0 detects the always false predicate in
resolver resulting in a FALSE condition which removes all other
predicates from the condition as well. This means, optimizer does not
have any keys to look into as there are no columns involved in the
condition. Since the keys are not setup, code path which marks the
inner table as const is not taken. In 5.7 however, since resolver does
not detect the always false condition, optimizer sets up keys to process
the other predicates that are part of the join condition. As the keys
are setup, range optimizer is called to estimate rowcount for the table
and it detects the always false condition. And after the detection,
inner table is marked as const.

Fix:
Even though we do not see the performance degradation that customer
reports, it would be nice to have this optimization even in 8.0. So
we now check for an ALWAYS FALSE join condition and mark the inner
table as const whenever applicable.

Changes in test results:
For an always false condition which does not have any other predicates
other than the always false predicate, on 5.7, optimizer does not mark
the inner table as const. However we do mark such tables as const
in 8.0 now.

Change-Id: I62f694f8b16662b49e06a4ffc15e3f3343ed1973