Bug#35115909: Wrong join result when one of the join columns is compared to a temporal literal Bug#35080094: MySQL server crash - Assertion CompatibleTypesForIndexLookup failed
This patch fixes two bugs:

1) Wrong results were seen in some case when one equality predicate
compared two string columns and another equality predicate compared
one of the string columns to a temporal constant. For example:

SELECT * FROM t WHERE a = b AND b = DATE'2023-01-01'

If a and b were VARCHAR columns with character set latin1, the
optimizer would propagate the DATE constant into the a = b predicate
and transform the query into:

SELECT * FROM t WHERE a = DATE'2023-01-01' AND b = DATE'2023-01-01'

The original a = b predicate would use string semantics for the
comparison, whereas the new a = DATE'2023-01-01' predicate uses DATE
semantics. The meaning of the query therefore changed, and wrong
results could be returned.

2) The hypergraph optimizer got an assertion failure when analyzing
which predicates were sargable. The assertion failure was seen if the
WHERE clause compared an indexed string column with a constant TIME
expression. It checked that all equality predicates that were part of
a multiple equality had arguments that were compatible for index
lookups.

Both of these issues were fixed by disabling creation of multiple
equalities from single equalities comparing a non-temporal string
column to a temporal constant.

Change-Id: Ibab64903eea83ce16ca4057154fa882a7e2b05f6