Bug#35118579 hypergraph ASAN crash in do_sub
The repro gives ASAN indication of buffer underflow if run with
hypergraph enabled.

We have a case of an implicitly grouped query with an ORDER BY using a
window function (two in repro). Since the query is implicitly grouped
the ORDER BY list is removed (result set has only one row, so no point
in sorting). And with the old optimizer, we do no windowing.

But with the hypergraph optimizer, windowing *is* performed.

Since the Query_object::order_list is emptied, we do not ever call
Item::split_sum_func on the ordering expression containing the window
function. This is bad, because it is what enables the window function
to be able to read columns from correct locations in tmp tables. Hence
we see the issue, when trying to evaluate Item_sum_sum::val_decimal:
we have a bad pointer, which is signalled by ASAN.

The patch solves this by removing the abandoned window functions from
Window::m_functions. This allows the hypergraph optimizer to omit the
windowing. This is achieved by allowing the loop calling
clean_up_after_removal (in Query_block::empty_order_list) to do its
thing also also for window functions, not just for grouped aggregates.

Change-Id: I3a5e00e3efa954aa3a98413ff90bbb3fb9e2de23