WL#15568: Periodic logging of Thread Pool connections
Problem: It is dfficult to determine the number and types of connections
established to a MySQL server which appears unresponsive in MDS.

Solution: Create a Thread Pool system variable which can be used to
enable periodic reports about the number of normal and admin connections
which are present and the number opened/closed since last report.
The logging is available only in cloud builds and enabled by default
with a report interval of 120 seconds.

Change-Id: I39daf88d3b8bd6a729695c403a0b18a212907b14