Bug#35129166 push_packet_overflow_warning(THD*, const char*):Assertion `is_nullable()' failed
Post-push fix. The test case fails when running with the hypergraph
optimizer because:

a) One of the queries has non-deterministic ordering (the expression
in the ORDER BY clause evaluates to 0 for all rows in the result set).
Fixed by adding --sorted_result.

b) One of the queries produces a warning when evaluating the
expression in the ORDER BY clause. The hypergraph optimizer elides the
ORDER BY clause completely and does not get a warning. Fixed by
disabling printing of warnings for the statement.

Change-Id: I49693dd33e45de5acc0e96c71ba3c8f31a74aa7f