Bug#35168702 xfrm_key_hash() should not be used for comparing keys
Fix incorrect usage of xfrm_key_hash() specific for 8.0

When the patch for Bug#30884622 was merged 7.5++ it likely encountered a conflict
due to the xfrm_key() / xfrm_attr() methods not existing in 8.0. However, 8.0
had the new methods xfrm_key_hash() intended *only* for extracting a binary string
from the key to later be Md5 hashed to create a hash value.

Unfortunately the calls to xfrm_key() were replaced with xfrm_key_hash() in the merge.
This opens up the possibility of two non-equal UCA-900 keys comparing as equal if the
extracted hash pattern turned out to be identical.

The intended replacement of ::xfrm_key() + memcmp() in 8.0 is to use the methods
cmp_key() and cmp_attr() instead. These methods does a character collation aware
compare if needed, often much more efficient than the previous usage of first
transforming the character keys: The String library strnncollsp() method is
use to do a direct collation aware compare, without the need for first
transforming the string.

Patch change as described above, which also significantly simplifies
the affected code block

Change-Id: I4b9129146eaa7945eae5d772b0f6ceb92c45e75f