Bug#35148945: String* Item::null_return_str(): Assertion `m_nullable' failed
The RANDOM_BYTES function is non-deterministic, but it is not marked
as such. This made the resolver think it was deterministic and
evaluate it. When it was evaluated again during execution, it returned
a different value, and the assumptions made by the resolver about the
value didn't hold. This led to assertion failures.

Fixed by marking the function as non-deterministic.

Change-Id: I9f185eb72fc3a67f50f80f4120da180bbd805ef5