WL#15517: [4] HeatWave Support for InnoDB Partitions - Support for Unloading Partitions
In this stage, we add functionality for unloading a set of partitions.
Unloading a partition is translated into deleting the corresponding rows
from the relation that exists in the rpdservers.

As UUs are organized per InnoDB partition id (lpid), instead of issuing
row-wise deletes, we use an algorithm that operates at the UU
granularity.

The flow for deleting a set of partitions is the following:
i) The user triggers an ALTER TABLE SECONDARY_UNLOAD PARTITION(p0,..,pn)
command
ii) The command is processed by UnloadTable in the plugin, but instead of
sending the regular unld_cmd to the cluster, we send a new command which
is called: LPART_UNLD_RPDODL_CMDT
iii) The rpdservers process this new command by running the algorithm
that deletes data in the UU granularity. Moreover, they persist the
command, so that it can be replayed upon recovery.
At this point, we note that we have modified the persistence semantics.
As both partition loads and unloads are treated as DMLs, we insert them
in the cp_list with change packets to be persisted, so that they are
persisted and be recovered if needed. Nevertheless, the partition
load/unload data are not of the same form with the update chunks that we
receive during Change Propagation. So, we generalize change packets, and
upon recovery a different handler is used based on the packet's type.
For partition unload commands, instead of data, we only have to persist
the command itself. Then, during recovery, we can use the same algorithm
that we use in the handler of LPART_UNLD_RPDODL_CMDT.
iii) After a successful deletion, the table remains both in GS and CP,
and only the partition metadata are updated. Also, after partition
unload, we do not issue a persistence GC.

~~~~~~~~~~~ Related Bugs ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Bug#36032836: Failed Condition vec->len_fipc_vec <= rapid::net::
kCommandChunkSize

Issue: The partition unload command takes as argument the list of
partition ids that should be unloaded. The size of this list can
exceed the max size we have for a rapid command.

Fix: Put a cap on the number of partitions that can be unloaded
at once. In the command size, we can fit up to 4k partitions.
However, as the partition unload command occupies the CP thread
for the entire time it is running, we have decided to set the
max number of partitions to unload at once to 1000.

~~~~~~~~~~~~

Bug#36009359: rpdserver log fills /var/rpdlogs with msg Primary key
overlap while inserting

Issue: In case the Btree has many pages and multiple levels, if the
key we delete is the last of a page, then rpdmutbt_lower_bound
function returns the lower bound that exists in that page and not in
the entire tree. Hence, our algorithm leaves some keys un-deleted
and the next time we try to reload the same partition, we get
primary key overlap errors.

Example:
PageA: 8a 8b 8c 6a 5c
PageB: 5d 4a 3c

If we delete 5c and run rpdmutbt_lower_bound, then, we are going to
get 6a, the algorithm will terminate and 5d will remain undeleted.

Fix: After removing an element, if the lower bound is not of the
same logical partition, then also try the next element

~~~~~~~~~~~~~~~~~

Bug#36161673: rpdserver silent crashes at rpdmutbt__pg_rem in WL15517

Issue: When removing an element from a btree page, the page may under
flow and we may have to merge it with the page to the left or to the
right. The merging function frees the right page. Hence, if we merge to
the left (and the current page is the right one), we free the current
page and then trying to access some of its data.

Fix: In case we free the current page, we should early return

Other fixes:
- When unloading a partition, move the stack allocation outside of the
loop: i) in case of large pkeys, we may hit stack overflow, ii) btree
keys for different superpartitions have the same size, and hence the
extra allocations are redundant

~~~~~~~~~~~~~~~~~~

Bug#36326757: rapid_reload_on_restart doesn't load partition table
after unload

Issue: Partition unload is not supposed to unload the table, but
only remove data from RAPID. However, as it shares the same
codepath with table unload, by default it unsets the secondary_load
flag. Hence, in case of mysld restart, the table is not reloaded

Fix: In case of partition unload, skip updating metadata in MySQL
side

Change-Id: Id42f2cbdc44d9797371297a0ce74f65343a7bfb7