Bug#35153587 Test programs use non public API via ndbclient
The ndbclient library is intended to provide the functionality of the
public NDBAPI and MGMAPI and the JNI-bindings for those.

But it also exposes many internal symbols and many test programs expect
some of those internal symbols from ndbclient library instead of using
the internal libraries.

Builds of test programs are changed to link against internal libraries
before the ndbclient library.

Also the enforcement of some internal symbols to be included in
ndbclient library are removed.

No change is made to not export internal symbols in ndbclient (shared)
library.

Change-Id: I1ac702a9b0fbd197c27010d9e734b11366c71ce4