BUG#34991101 - Memory leak on Network_provider_manager::open_xcom_connection()
Description
==================================
There is a leak when running the test
gr_add_node_with_ill_configurated_stack with the server built
with ASAN.

Analysis and Fix
==================================
The leak occurs when we try to connect with the seeds to join the
group and this server is stopped in the process. When that connection
descriptor returns and is non null, it must be destroyed.

Change-Id: I2d86af1b1630beac04279f66c03f0a998c2941cc