Bug#35163463 explicitly wait for connection pool to be filled
The tests which check if warnings survive connection sharing assume:

1. a statement with warnings (like DO 0/0)
2. connection is pooled
3. SHOW WARNINGS

Afterwards the test wants to close the pooled connections and check if
the warnings can still be read, even though the server connections are
closed.

1. a statement with warnings (like DO 0/0)
  - router sends a SHOW WARNINGS and captures the result
2. connection is pooled
3. SHOW WARNINGS

Current Implementation:

To ensure the connection is pooled, the test currently relies on
sending an PING after the statement-with-warnings.

1. a statement with warnings (like DO 0/0)
  - router sends a SHOW WARNINGS and captures the result
2. PING
3. connection is pooled
4. SHOW WARNINGS

... as the PING has no warnings and the connection would be pooled the
PING's OK is returned to the client.

That assumption breaks as soon as the PING returns a warning itself,
which would overwrite the warnings of the "statement-with-warnings".

Change
------

- replaced the PING by a active check if the connection pool has a
  connection.
- replaced a 50ms sleep with an active check if the connection pool is
  empty.

The test now runs 4 seconds faster as it sleeps less.

  74.3s -> 70.5sec.

Change-Id: I4b98ffd2bc56509e5f7a3fbd89de7b149dbe7c34