Bug#35168639 Crash in Item_ref::walk during cast injection Bug#35204224 Crash with unused window function and order by
The first bug is a a regression due to patch for Bug#35118579 which
removed unused window functions after an ORDER BY elimination. That
patch left window functions in the select list, which isn't safe, as
evidenced by this patch.

The fix is to replace the hidden expression in the select list
containing the window function with a (hidden) Item_null. We replace
rather than shrink the list to avoid upsetting assumptions on indexes
on/size of Query_object::fields and base_ref_items.

The second is another variant: a view reference (well, really an
Item_view_ref representing a derived table reference) in the ORDER BY
list is not removed because of the presence of a window (not a window
function in this case, just a window definition): so we do not chop
off the ordering expression from the select list in empty_order_list,
which leaves the hidden reference (f2) still in the select
list. Later, the derived table is merged, and it is concluded that f2
is no longer in use since its reference count has fallen to zero (in
QB:: delete_unused_merged_columns which nulls it out with
"transl->item = nullptr;"), so when trying to optimize f2 (it is
still in the select list as a hidden item), we see the crash trying to
access the ref's m_ref_item which now points to a nullptr.

The fix is to *always* replace hidden order by expressions from the
select list with a null in the presence of windows on the query block,
i.e. not only if the expression contains window function calls.

Side effect from second fix: the following query used to fail with
subquery_to_derived=on, but now succeeds:

SELECT SUM(f1), ROW_NUMBER() OVER (),
       (SELECT MIN(f1) FROM t1)
FROM t1 ORDER BY f1

since f1 is removed. It used to succeed with subquery_to_derived=off
and still does.

Change-Id: I6d4b472ad4e65cd91fe6dc9515e5a6c92dd3a3c0